QuatschDjango/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.2
django-braces==1.15.0
django-cors-headers==3.14.0
django-markdownx==4.0.2
django-oauth-toolkit==2.2.0
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.14.0
filelock==3.12.0
fsspec==2023.5.0
huggingface-hub==0.14.1
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
jwcrypto==1.4.2
Markdown==3.4.3
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
pypng==0.20220715.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
regex==2023.5.5
requests==2.28.2
requests-oauthlib==1.3.1
sacremoses==0.0.53
sentencepiece==0.1.99
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.29.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wrapt==1.15.0
